perf: implement encodeRegenerate
Also separate encode-ranges from decode-ranges. The encoding utility is not used by end-users.
