+++ static/decode-ranges.js
@@ -1,3 +1,8 @@
+/**
+ * This file ships to end users.
+ * See scripts/encode-ranges.js for encode utilities
+ */
+
 const base64enc =
 	'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_';
 
@@ -55,26 +60,6 @@ function decodeDeltas(input) {
 	return output;
 }
 
-/**
-  * Base64 encode variable-length deltas (5/10/15/21-bit).
-  */
-function encodeDeltas(input) {
-	const output = [];
-	for (let i = 0; i < input.length; ++i) {
-		const x = input[i];
-		if ((x >> 5) === 0) {
-			output.push(x);
-		} else if ((x >> 10) === 0) {
-			output.push(32 + (x >> 6), x);
-		} else if ((x >> 15) === 0) {
-			output.push(48 + (x >> 12), x >> 6, x);
-		} else {
-			console.assert((x >> 21) === 0, `delta ${x} out of range`);
-			output.push(56 + (x >> 18), x >> 12, x >> 6, x);
-		}
-	}
-	return output.map(x => base64enc[x & 63]).join('');
-}
 
 /**
   * RLE + base64 decode code point ranges.
@@ -91,24 +76,4 @@ function decodeRanges(input) {
 	return ranges;
 }
 
-/**
-  * RLE + base64 encode code point ranges.
-  */
-function encodeRanges(values) {
-	const deltas = [];
-	for (let end = -1, i = 0; i < values.length; ) {
-		const begin = values[i];
-		console.assert(begin > end, `code point ${begin} out of order`);
-		deltas.push(begin - end - 1);
-		end = begin + 1;
-		while (++i < values.length && values[i] === end) {
-			++end;
-		}
-		deltas.push(end - begin - 1);
-	}
-	return encodeDeltas(deltas);
-}
-
-module.exports = Object.defineProperty(
-	decodeRanges, 'encode', { value: encodeRanges }
-);
+module.exports = decodeRanges;
